requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
faiss-cpu
sentence-transformers
pillow
